src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
click==8.1.8
    # via uvicorn
fastapi==0.115.8
    # via image-blur (pyproject.toml)
h11==0.14.0
    # via uvicorn
idna==3.10
    # via anyio
jinja2==3.1.5
    # via image-blur (pyproject.toml)
mangum==0.19.0
    # via image-blur (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
pillow==11.1.0
    # via image-blur (pyproject.toml)
pip==25.0
    # via image-blur (pyproject.toml)
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
python-multipart==0.0.20
    # via image-blur (pyproject.toml)
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   mangum
    #   pydantic
    #   pydantic-core
uvicorn==0.34.0
    # via image-blur (pyproject.toml)
